+++ src/app/cuenta/cuenta.page.scss
@@ -23,3 +23,8 @@
   margin-left: 20px;
   color: hsl(0, 71%, 59%);
 }
+
+.per {
+  margin-left: 20px;
+  color: white;
+}
