+++ src/app/pages/menu/menu.page.html
@@ -11,9 +11,6 @@
 </ion-header>
 
 <ion-content>
-  <ion-button color="light" expand="full" [routerLink]="['/perfil']" ><span class="material-symbols-outlined">
-    smart_display
-    </span> Videos</ion-button>
   <h3>Beber agua</h3>
   <ion-card>
     <ion-button class="bntn" fill="clear" id="present-alert"><span class="material-symbols-outlined">
