+++ src/app/perfil/perfil.page.html
@@ -16,4 +16,9 @@
       <p>Toca para cambiar la foto</p>
     </ion-text>
   </div>
+  <ion-item>
+    <ion-label position="floating">Cambiar nombre de usuario</ion-label>
+    <ion-input [(ngModel)]="newUserName"></ion-input>
+  </ion-item>
+  <ion-button expand="full" (click)="updateUserName()">Actualizar Usuario</ion-button>
 </ion-content>
