+++ src/app/perfil/perfil.page.scss
@@ -1,0 +1,18 @@
+.profile-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin: 2rem auto;
+  text-align: center;
+
+  .profile-image {
+    width: 150px;
+    height: 150px;
+    margin-bottom: 1rem;
+  }
+
+  .change-photo-text {
+    margin-top: 0.5rem;
+  }
+}
