+++ src/app/perfil/perfil.page.ts
@@ -3,6 +3,7 @@ import { ActionSheetController } from '@ionic/angular';
 import { ImageService } from 'src/service/image.service';
 import { FirebaseStorageService } from 'src/service/firebase-storage.service';
 import { UpdateUserImageUseCase } from 'src/use-cases/updateimage-user.usecase';
+import { UpdateUserNameUseCase } from 'src/use-cases/updateName-user.usecase';
 
 @Component({
   selector: 'app-perfil',
@@ -13,16 +14,39 @@ export class PerfilPage implements OnInit {
   currentImage: string = '';
   userName: string = '';
   userId: string = '';
+  currentUserName: string = '';
+  newUserName: string = '';
+
 
   constructor(
     private actionSheetController: ActionSheetController,
     private imageService: ImageService,
     private storageService: FirebaseStorageService,
     private UpdateUserImageUseCase: UpdateUserImageUseCase,
+    private updateUserNameUC: UpdateUserNameUseCase,
   ) {}
 
   async ngOnInit() {
     await this.loadUserProfile();
+    await this.loadUserName();
+  }
+
+  async loadUserName() {
+    this.currentUserName = (await this.updateUserNameUC.getUserName()) ?? '';
+    this.newUserName = this.currentUserName || '';
+  }
+
+  async updateUserName() {
+    try {
+      const result = await this.updateUserNameUC.updateUserName(this.newUserName);
+      if (result) {
+        console.log('Nombre de usuario actualizado correctamente');
+      } else {
+        console.error('Error al actualizar el nombre de usuario');
+      }
+    } catch (error) {
+      console.error('Error al actualizar el nombre de usuario:', error);
+    }
   }
 
   async loadUserProfile() {
@@ -70,6 +94,14 @@ export class PerfilPage implements OnInit {
 
   private async updateProfilePhoto(imageUrl: string) {
     try {
+      if (this.currentImage && this.currentImage.includes('firebase')) {
+        try {
+          const oldImagePath = this.currentImage.split('user-image%2F')[1].split('?')[0];
+          await this.storageService.deleteImage('user-image/' + oldImagePath);
+        } catch (deleteError) {
+          console.warn('No se pudo eliminar la imagen anterior:', deleteError);
+        }
+      }
       const uploadPath = `user-image/${this.userId}-${new Date().getTime()}.png`;
       const uploadedUrl = await this.storageService.uploadImage(uploadPath, imageUrl);
